Hackathon_1/requirements.txt
# ROS 2 dependencies (typically installed separately, but listed for reference)
# ros-humble-ros2launch
# ros-humble-nav2
# ros-humble-vision-opencv

# Core dependencies
openai>=1.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.21.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
pyyaml>=6.0

# Audio processing
pyaudio>=0.2.11
soundfile>=0.12.0
librosa>=0.10.0

# ROS 2 Python interface (install via apt: python3-colcon-common-extensions ros-humble-rclpy)
rclpy>=3.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Linting and formatting
flake8>=6.0.0
black>=23.0.0
isort>=5.0.0
mypy>=1.0.0

# Environment management
python-dotenv>=1.0.0